fix: alter password in connection string.
This is to hide the password for READ-WRITE access to the database in order to protect it from others who may accidentally or maliciously perform corrosive actions to it.

Whilst can track back to previous versions of this file to get the old password, the password has now been changed so the old password will no longer give READ-WRITE access.