fix: change reference use of key.
This is to successfully import countries data into several collections, dynamically.

Problem before was due to the key from the dictionary not being treated as variable but as a string. Therefore needed to change the way we refer to this key by referring to it as a string, db[key] -> e.g. db['names'], rather than as a variable, db.key -> db.'names'.